[K/N] custom-alloc: fix nits
Change all generic pointer types to uint8_t*
Reformat code with git-clang-format
Make constants into constexpr in shared file

+++ kotlin-native/runtime/src/custom_alloc/cpp/Cell.cpp
@@ -14,9 +14,8 @@ Cell::Cell(uint32_t size) noexcept : isAllocated_(false), size_(size) {
     CustomAllocDebug("Cell@%p::Cell(%u)", this, size);
 }
 
-uint64_t* Cell::TryAllocate(uint32_t cellsNeeded) noexcept {
-    CustomAllocDebug("Cell@%p{ allocated = %d, size = %u }::TryAllocate(%u)",
-            this, isAllocated_, size_, cellsNeeded);
+uint8_t* Cell::TryAllocate(uint32_t cellsNeeded) noexcept {
+    CustomAllocDebug("Cell@%p{ allocated = %d, size = %u }::TryAllocate(%u)", this, isAllocated_, size_, cellsNeeded);
     if (isAllocated_ || cellsNeeded > size_) {
         CustomAllocDebug("Failed to allocate in Cell");
         return nullptr;
@@ -27,18 +26,16 @@ uint64_t* Cell::TryAllocate(uint32_t cellsNeeded) noexcept {
     size_ = remainingSize;
     newBlock->isAllocated_ = true;
     newBlock->size_ = cellsNeeded;
-    RuntimeAssert(remainingSize == 0 || size_ + newBlock->size_ == oldSize,
-            "sizes don't add up");
+    RuntimeAssert(remainingSize == 0 || size_ + newBlock->size_ == oldSize, "sizes don't add up");
     return newBlock->Data(); // Payload starts after header
 }
 
-uint64_t* Cell::Data() noexcept {
-    return reinterpret_cast<uint64_t*>(this+1);
+uint8_t* Cell::Data() noexcept {
+    return reinterpret_cast<uint8_t*>(this + 1);
 }
 
 void Cell::Deallocate() noexcept {
-    CustomAllocDebug("Cell@%p{ allocated = %d, size = %u }::Deallocate()",
-            this, isAllocated_, size_);
+    CustomAllocDebug("Cell@%p{ allocated = %d, size = %u }::Deallocate()", this, isAllocated_, size_);
     RuntimeAssert(isAllocated_, "Cell is not currently allocated");
     isAllocated_ = false;
 }
@@ -47,4 +44,4 @@ Cell* Cell::Next() noexcept {
     return this + size_;
 }
 
-}  // namespace kotlin::alloc
+} // namespace kotlin::alloc
